chore: unify jumpToMessage
