+++ apps/meteor/client/components/message/toolbar/useJumpToMessageContextAction.tsx
@@ -1,25 +1,33 @@
 import type { IMessage } from '@rocket.chat/core-typings';
 import { useEffect } from 'react';
 
+import type { MessageActionContext } from '../../../../app/ui-utils/client/lib/MessageAction';
 import { MessageAction } from '../../../../app/ui-utils/client/lib/MessageAction';
 import { setMessageJumpQueryStringParameter } from '../../../lib/utils/setMessageJumpQueryStringParameter';
 
-export const useJumpToSearchMessageAction = (message: IMessage) => {
+export const useJumpToMessageContextAction = (
+	message: IMessage,
+	{ id, order, hidden, context }: { id: string; order: number; hidden?: boolean; context: MessageActionContext[] },
+) => {
 	useEffect(() => {
+		if (hidden) {
+			return;
+		}
+
 		MessageAction.addButton({
-			id: 'jump-to-search-message',
+			id,
 			icon: 'jump',
 			label: 'Jump_to_message',
-			context: ['search'],
+			context,
 			async action() {
 				setMessageJumpQueryStringParameter(message._id);
 			},
-			order: 100,
+			order,
 			group: 'message',
 		});
 
 		return () => {
-			MessageAction.removeButton('jump-to-search-message');
+			MessageAction.removeButton(id);
 		};
-	}, [message._id]);
+	}, [hidden, context, id, message._id, order]);
 };
